both sensor values for light and for temperature are continously sent to the coordinator

+++ firmware-wsn/include/messaging.h
@@ -20,7 +20,7 @@ extern HAL_UsartDescriptor_t usart;
 
 void forwardMessageToPC();
 void createBeaconList();
-void usartRcvd(uint8_t size);
+void usartRcvd(uint16_t size);
 void usart_Init();
 void sendUart(uint8_t* string, uint8_t size);
 void assembleBeaconMessage();
